pyproject.toml
[project]
name = "guardrails_integration"
version = "0.1.0"
description = "REST API for Guardrails integration (FastAPI + Guardrails Hub)"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111.0",
    "guardrails-ai>=0.4.7",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
    "jsonschema>=4.22.0",
    "jmespath>=1.0.1",
    "pytest>=8.2.0",
]